# Construct kerning/mark lookup builders direclty