This is followup to #598; since we don't need access to the feature builder in order to construct anchors or value records, we can now construct the various builders directly in the relevant tasks.