@rsheeter 2023-11-29
Btw does hyperfine show any improvement on this vs main?

```shell
$ cargo build --release && hyperfine --warmup 25 --runs 100 --prepare 'rm -rf build/' 'target/release/fontc ../OswaldFont/sources/Oswald.glyphs'

$ cargo build --release && hyperfine --warmup 5 --prepare 'rm -rf build/' 'target/release/fontc ../googlesans/source/GoogleSans/GoogleSans.designspace'
````