@cmyr 2023-11-29
before:

```
Benchmark 1: target/release/fontc ../fontfiles/googlesans/source/GoogleSans/GoogleSans.designspace
  Time (mean ± σ):      4.349 s ±  0.244 s    [User: 7.100 s, System: 3.082 s]
  Range (min … max):    4.152 s …  4.919 s    10 runs
```

after:

```
Benchmark 1: target/release/fontc ../fontfiles/googlesans/source/GoogleSans/GoogleSans.designspace
  Time (mean ± σ):      4.260 s ±  0.052 s    [User: 7.196 s, System: 3.029 s]
  Range (min … max):    4.188 s …  4.368 s    10 runs
```

so nothing super significant.

This shouldn't really surprise us, since we're not saving ourselves a ton of work here, just removing one extra iteration/allocation pass over all the mark/kern rules.