@rsheeter 2023-11-29
Probably slightly faster, seems about right. Ty for checking.